MB-73781: Fix stats aggregation
Reviewed-on: https://review.couchbase.org/c/eventing/+/252774
Tested-by: Eventing CI Bot
(cherry picked from commit decdf573d680c008f64fc552c1ab11fe320ee7b2)
Reviewed-on: https://review.couchbase.org/c/eventing/+/253392
Well-Formed: Restriction Checker

+++ function_manager/function_handler/function_handler_impl.go
@@ -803,8 +803,8 @@ func (fHandler *funcHandler) ReceiveMessage(msg *processManager.ResponseMessage)
 		case processManager.StatsAckBytes:
 			vbHandler := fHandler.vbHandler.Load()
 			unackedCount, unackedBytes := vbHandler.AckMessages(msg.Value)
-			fHandler.statsHandler.AddExecutionStats("agg_queue_memory", unackedBytes)
-			fHandler.statsHandler.AddExecutionStats("agg_queue_size", unackedCount)
+			fHandler.statsHandler.AddExecutionStats("agg_queue_memory", float64(unackedBytes))
+			fHandler.statsHandler.AddExecutionStats("agg_queue_size", float64(unackedCount))
 
 		default:
 			fHandler.statsHandler.handleStats(msg)
